rename composable demos and remove deprecated window scroller demo
rename composable window scroller demo
